User adjustments

+++ DomesticAppliencesForum/src/main/java/com/telerikacademy/domesticappliencesforum/services/PostServiceImpl.java
@@ -1,5 +1,6 @@
 package com.telerikacademy.domesticappliencesforum.services;
 
+import com.telerikacademy.domesticappliencesforum.exceptions.UnauthorizedOperationException;
 import com.telerikacademy.domesticappliencesforum.models.Post;
 import com.telerikacademy.domesticappliencesforum.models.User;
 import com.telerikacademy.domesticappliencesforum.repositories.PostRepository;
@@ -8,22 +9,22 @@ import com.telerikacademy.domesticappliencesforum.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
 public class PostServiceImpl implements PostService {
 
     private PostRepository postRepository;
+    private UserRepository userRepository;
 
     @Autowired
-    public PostServiceImpl() {
-        this.postRepository = new PostRepositoryImpl();
+    public PostServiceImpl(UserRepository userRepository) {
+        this.postRepository = new PostRepositoryImpl(userRepository);
 
     }
 
     @Override
-    public List<Post> getAllPosts(String title, Integer authorId, String localDate) {
+    public List<Post> getAllPosts(String title, User authorId, String localDate) {
         return postRepository.getAllPosts(title, authorId, localDate);
     }
 
@@ -39,13 +40,20 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void modify(Post post) {
+    public void modify(Post post, User user) {
+        if (post.getAuthorId().equals(user.getId())){
+            throw new UnauthorizedOperationException("Admins cannot modify posts");
+        }
         //TODO Да направим верификация за потребител да може да променя само неговите си постове
         postRepository.modify(post);
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, User user) {
+        Post post = postRepository.getPostById(id);
+        if(!user.isAdmin() || !post.getAuthorId().equals(user.getId())){
+            throw new UnauthorizedOperationException("Only admins can delete");
+        }
         //TODO Да направим верификация за потребител да може да променя само
         // неговите си постове и админа - всички
         postRepository.delete(id);
